add share button on the video player view
fix #205
