+++ lib/models/baseVideo.dart
@@ -1,6 +1,8 @@
+import 'db/server.dart';
 import 'imageObject.dart';
+import 'interfaces/sharelink.dart';
 
-abstract class BaseVideo{
+abstract class BaseVideo implements ShareLinks {
   String title;
   String videoId;
   int lengthSeconds;
@@ -10,4 +12,14 @@ abstract class BaseVideo{
   List<ImageObject> videoThumbnails;
 
   BaseVideo(this.title, this.videoId, this.lengthSeconds, this.author, this.authorId, this.authorUrl, this.videoThumbnails);
+
+  @override
+  String getInvidiousLink(Server server) {
+    return '${server.url}/watch?v=$videoId';
+  }
+
+  @override
+  String getYoutubeLink() {
+    return 'https://www.youtube.com/watch?v=$videoId';
+  }
 }
